requirements.txt
psycopg[binary] == 3.2.3
python-dotenv == 1.0.1
fastapi==0.115.6
streamlit==1.40.2